add test for bug
