+++ portal/tests/test_aimmo_dashboards.py
@@ -1,56 +1,10 @@
-from collections import namedtuple
-
 import pytest
 from aimmo.models import Game
 from common.models import Class
-from common.tests.utils.classes import create_class_directly
-from common.tests.utils.student import (
-    create_school_student_directly,
-    create_independent_student_directly,
-)
-from common.tests.utils.teacher import signup_teacher_directly
 from django.test.client import Client
 from django.urls.base import reverse
 
-from portal.tests.utils.organisation import create_organisation_directly
-from .utils.aimmo_games import create_aimmo_game_directly
-from .utils.worksheets import create_worksheet_directly
-
-SchoolStudent = namedtuple("student", ["username", "password"])
-IndependentStudent = namedtuple("independent_student", ["username", "password"])
-
-
-@pytest.fixture
-def teacher1_email(db) -> str:
-    teacher_email, _ = signup_teacher_directly()
-    return teacher_email
-
-
-@pytest.fixture
-def class1(db, teacher1_email) -> Class:
-    create_organisation_directly(teacher1_email)
-    klass, _, _ = create_class_directly(teacher1_email)
-    return klass
-
-
-@pytest.fixture
-def student1(db, class1) -> SchoolStudent:
-    username, password, _ = create_school_student_directly(class1.access_code)
-    return SchoolStudent(username, password)
-
-
-@pytest.fixture
-def independent_student1(db) -> IndependentStudent:
-    username, password, _ = create_independent_student_directly()
-    return IndependentStudent(username, password)
-
-
-@pytest.fixture
-def aimmo_game1(db, class1) -> Game:
-    worksheet = create_worksheet_directly()
-    worksheet.student_pdf_name = "TestPDFName"
-    worksheet.save()
-    return create_aimmo_game_directly(class1, worksheet)
+from .conftest import IndependentStudent, SchoolStudent
 
 
 @pytest.mark.django_db
