+++ portal/tests/test_materials.py
@@ -34,10 +34,12 @@
 # copyright notice and these terms. You must not misrepresent the origins of this
 # program; modified versions of the program must be marked as such and not
 # identified as the original program.
-from django.test import TestCase
-
-from portal.templatetags.table_tags import resource_sheets_table, lengthen_list
+from django.test import Client, TestCase
+from django.urls.base import reverse
+from portal.templatetags.table_tags import lengthen_list, resource_sheets_table
 from portal.views.materials_viewer import get_links
+
+from .conftest import TeacherLoginDetails
 
 
 class MaterialsTests(TestCase):
@@ -73,3 +75,14 @@ class MaterialsTests(TestCase):
         pdf_name = "KS1_session_3"
         result = get_links(pdf_name)
         assert result == [("KS1_S3_1", "KS1 S3 1")]
+
+
+def test_materials_viewer_redirect(
+    live_server, client: Client, teacher1: TeacherLoginDetails
+):
+    client.login(email=teacher1.email, password=teacher1.password)
+    response = client.get(
+        reverse("materials_viewer_redirect", kwargs={"pdf_name": "KS1_S3_1"}),
+        follow=True,
+    )
+    assert response.status_code == 200
